Removing mpsitr.c, a relic of custom/cet, and merging its contents into mpsi.c.  Reponse to review <https://github.com/Ravenbrook/mps/pull/214#issuecomment-1590952221>.
