+++ code/commpre.nmk
@@ -147,7 +147,6 @@ MPMCOMMON=\
     [meter] \
     [mpm] \
     [mpsi] \
-    [mpsitr] \
     [nailboard] \
     [policy] \
     [pool] \
