+++ code/mpsi.c
@@ -1,7 +1,7 @@
 /* mpsi.c: MEMORY POOL SYSTEM C INTERFACE LAYER
  *
  * $Id$
- * Copyright (c) 2001-2020 Ravenbrook Limited.  See end of file for license.
+ * Copyright (c) 2001-2023 Ravenbrook Limited.  See end of file for license.
  * Portions copyright (c) 2002 Global Graphics Software.
  *
  * .purpose: This code bridges between the MPS interface to C,
@@ -47,6 +47,7 @@
 #include "mpm.h"
 #include "mps.h"
 #include "sac.h"
+#include "trans.h"
 
 #include <stdarg.h>
 
@@ -2081,9 +2082,87 @@ void _mps_args_set_key(mps_arg_s args[MPS_ARGS_MAX], unsigned i,
 }
 
 
+/* Transforms */
+
+
+mps_res_t mps_transform_create(mps_transform_t *mps_transform_o,
+                               mps_arena_t arena)
+{
+  Transform transform = NULL;
+  Res res;
+
+  AVER(mps_transform_o != NULL);
+
+  ArenaEnter(arena);
+  res = TransformCreate(&transform, arena);
+  ArenaLeave(arena);
+  if (res != ResOK)
+    return res;
+
+  *mps_transform_o = (mps_transform_t)transform;
+  return MPS_RES_OK;
+}
+
+
+mps_res_t mps_transform_add_oldnew(mps_transform_t transform,
+                                   mps_addr_t *mps_old_list,
+                                   mps_addr_t *mps_new_list,
+                                   size_t mps_count)
+{
+  Ref *old_list = (Ref *)mps_old_list;
+  Ref *new_list = (Ref *)mps_new_list;
+  Count count = mps_count;
+  Arena arena;
+  Res res;
+
+  AVER(mps_old_list != NULL);
+  AVER(mps_new_list != NULL);
+  /* count: cannot check */
+
+  arena = TransformArena(transform);
+
+  ArenaEnter(arena);
+  res = TransformAddOldNew(transform, old_list, new_list, count);
+  ArenaLeave(arena);
+
+  return res;
+}
+
+
+mps_res_t mps_transform_apply(mps_bool_t *applied_o,
+                              mps_transform_t transform)
+{
+  Arena arena;
+  Res res;
+
+  AVER(applied_o != NULL);
+
+  arena = TransformArena(transform);
+  ArenaEnter(arena);
+  STACK_CONTEXT_BEGIN(arena) {
+    res = TransformApply(applied_o, transform);
+  } STACK_CONTEXT_END(arena);
+  ArenaLeave(arena);
+
+  return res;
+}
+
+
+void mps_transform_destroy(mps_transform_t transform)
+{
+  Arena arena;
+
+  arena = TransformArena(transform);
+
+  ArenaEnter(arena);
+  TransformDestroy(transform);
+  ArenaLeave(arena);
+}
+
+
 /* C. COPYRIGHT AND LICENSE
  *
- * Copyright (C) 2001-2020 Ravenbrook Limited <https://www.ravenbrook.com/>.
+ * Copyright (C) 2001-2023 Ravenbrook Limited <https://www.ravenbrook.com/>.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
